Build movie detail page
For the movie detail, there will be 3 sections.
First is the header section which will display the movie image and short description.
Second is the movie detail which includes the release data, runtime, genre.
Last section is the list of casts

Each of this section requires fetching data from API.