Build the WatchList tab page and fix some layout issues
Build the WatchList page and provide some dummy text for now while waiting for the core data functionality to be fully set up. Fix the layout issues with the dynamic heights of table view. Add extra fields in the core data model. At the minimum, it should have the data for the movie id, image, title and year.